backend/requirements.txt
Django==4.2.1
djangorestframework==3.14.0
python-dotenv==1.0.0
django-filter==23.2
psycopg2_binary==2.9.6
django-colorfield==0.8.0
djoser==2.2.0
drf-extra-fields==3.5.0
drf-yasg==1.21.5
django_rest_swagger==2.2.0
gunicorn==20.1.0
sorl-thumbnail